+++ packages/wiz-ui-next/src/components/base/inputs/selectbox/types.ts
@@ -1,6 +1,6 @@
-export interface SelectBoxOption {
+export interface SelectBoxOption<T = number> {
   label: string;
   exLabel?: string;
-  value: number;
+  value: T;
   disabled?: boolean;
 }